add interface declarations in separate files
