+++ src/clients/http.client.ts
@@ -1,13 +1,6 @@
 import axios, { AxiosError, AxiosInstance, AxiosRequestConfig } from 'axios';
 import axiosRetry from 'axios-retry';
-
-export interface IHttpClient {
-  get<T>(url: string, params?: Record<string, unknown>): Promise<T>;
-  post<T>(url: string, body?: unknown, params?: Record<string, unknown>): Promise<T>;
-  put<T>(url: string, body?: unknown, params?: Record<string, unknown>): Promise<T>;
-  patch<T>(url: string, body?: unknown, params?: Record<string, unknown>): Promise<T>;
-  delete<T>(url: string, params?: Record<string, unknown>): Promise<T>;
-}
+import { IHttpClient } from '../interfaces/http.client.interface';
 
 export type ErrorMapper<E = unknown> = (error: AxiosError<E>) => Error;
