+++ src/config/di.tokens.ts
@@ -7,8 +7,8 @@ export const TOKENS = {
   WeatherService: Symbol.for('WeatherService'),
   WeatherController: Symbol.for('WeatherController'),
   WeatherDigestJob: Symbol.for('WeatherDigestJob'),
-  IMailTransport: Symbol.for('IMailTransport'),
-  IMailService: Symbol.for('IMailService'),
+  IEmailTransport: Symbol.for('IEmailTransport'),
+  IEmailService: Symbol.for('IEmailService'),
   SubscriptionService: Symbol.for('SubscriptionService'),
   SubscriptionController: Symbol.for('SubscriptionController'),
   ISubscriptionRepository: Symbol.for('ISubscriptionRepository'),
