+++ src/enums/subscription-frequency.enum.ts
@@ -1,0 +1,4 @@
+export enum SubscriptionFrequencyEnum {
+  HOURLY = 'hourly',
+  DAILY = 'daily',
+}
