+++ src/interfaces/cache.service.interface.ts
@@ -1,0 +1,4 @@
+export interface ICacheService<T> {
+  get(key: string): Promise<T | null>;
+  set(key: string, value: T, ttlSeconds: number): Promise<void>;
+}
