+++ src/interfaces/email-send-options.interfaces.ts
@@ -1,4 +1,4 @@
-export interface SendOptions {
+export interface IEmailSendOptions {
   to: string;
   subject: string;
   html: string;
