+++ src/interfaces/email.client.interface.ts
@@ -1,0 +1,6 @@
+import type SMTPTransport from 'nodemailer/lib/smtp-transport';
+
+export interface IEmailTransport {
+  verify(): Promise<true>;
+  send(options: SMTPTransport.Options & { from: string }): Promise<unknown>;
+}
