+++ src/interfaces/email.service.interface.ts
@@ -1,0 +1,5 @@
+import { IEmailSendOptions } from './email-send-options.interfaces';
+
+export interface IEmailService {
+  send(opts: IEmailSendOptions): Promise<void>;
+}
