+++ src/interfaces/logger.service.interface.ts
@@ -1,0 +1,5 @@
+export interface ILogger {
+  info(message: string): void;
+  warn(message: string): void;
+  error(message: string, error?: unknown): void;
+}
