+++ src/interfaces/weather-api-error-response.interface.ts
@@ -1,0 +1,6 @@
+export interface IWeatherApiErrorData {
+  error: {
+    code: number;
+    message: string;
+  };
+}
