+++ src/interfaces/weather-api-response.interface.ts
@@ -1,0 +1,13 @@
+export interface IWeatherApiCondition {
+  text: string;
+}
+
+export interface IWeatherApiCurrent {
+  temp_c: number;
+  humidity: number;
+  condition: IWeatherApiCondition;
+}
+
+export interface IWeatherApiResponse {
+  current: IWeatherApiCurrent;
+}
