+++ src/interfaces/weather-api.client.interface.ts
@@ -1,0 +1,5 @@
+import { IWeatherApiResponse } from '../interfaces/weather-api-response.interface';
+
+export interface IWeatherApiClient {
+  fetchCurrent(city: string): Promise<IWeatherApiResponse>;
+}
