+++ src/interfaces/weather.mapper.interface.ts
@@ -1,0 +1,6 @@
+import { WeatherDto } from '../dto/weather.dto';
+import { IWeatherApiResponse } from '../interfaces/weather-api-response.interface';
+
+export interface IWeatherMapper {
+  mapCurrentWeather(raw: IWeatherApiResponse): WeatherDto;
+}
