+++ src/repositories/subscription.repository.ts
@@ -1,17 +1,9 @@
 import { AppDataSource } from '../config/dataSource';
 import { Subscription, Frequency } from '../models/subscription.entity';
 import { FindOptionsWhere } from 'typeorm';
+import { ISubscriptionRepository } from '../interfaces/subscription.repository.interface';
 
 type SubscriptionField = 'confirmation_token' | 'unsubscribe_token';
-
-export interface ISubscriptionRepository {
-  findByToken(token: string, field: SubscriptionField): Promise<Subscription | null>;
-  findExisting(email: string, city: string, frequency: Frequency): Promise<Subscription | null>;
-  save(sub: Partial<Subscription>): Promise<Subscription>;
-  confirm(sub: Subscription): Promise<void>;
-  remove(sub: Subscription): Promise<void>;
-  findConfirmedByFrequency(frequency: Frequency): Promise<Subscription[]>;
-}
 
 export class SubscriptionRepository implements ISubscriptionRepository {
   private repo = AppDataSource.getRepository(Subscription);
